Add files via upload
This project aims to train and test a machine-learning model that can accurately predict whether a song is going to be popular or not. In order to achieve this, we will be testing out different machine learning models and will look at what attributes of a song are the most important for determining its popularity. It is implemented using R.